requirements.txt
defusedxml==0.7.1
protobuf==3.15.8
pycryptodomex==3.10.1
requests==2.25.1
Pillow==8.2.0
tqdm==4.60.0
mutagen==1.45.1
ffmpeg-python==0.2.0
